feat(deps): add elevenlabs extra and refactor live tests
Add `elevenlabs` to the optional dependencies in `pyproject.toml`.
Relocate `SoundDevicePlayer` tests from `test/live/test_cascade.py` to a
new dedicated test file `test/live/test_sound_device.py` to improve
test organization and dependency management.
